+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed issue in `CMakePresets.json` where Ninja presets were broken
+- Fixed io profiler report format
 
 ### Added